Improve e2e test (#270)

+++ test/e2e/api_server_test.go
@@ -24,23 +24,33 @@ var _ = Describe("API server", func() {
 
 	BeforeEach(func() {
 		f = root.Invoke()
-		os.RemoveAll(filepath.Join("/tmp", "indices"))
+		if !f.SelfHostedOperator {
+			os.RemoveAll(filepath.Join("/tmp", "indices"))
+		}
 	})
 
 	JustBeforeEach(func() {
-		By("Starting Kubed")
-		stopCh = make(chan struct{})
-		err := f.RunKubed(stopCh, clusterConfig)
-		Expect(err).NotTo(HaveOccurred())
+		if f.SelfHostedOperator {
+			By("Restarting kubed operator")
+			err := f.RestartKubedOperator(&clusterConfig)
+			Expect(err).NotTo(HaveOccurred())
+		} else {
+			By("Starting Kubed")
+			stopCh = make(chan struct{})
+			err := f.RunKubed(stopCh, clusterConfig)
+			Expect(err).NotTo(HaveOccurred())
 
-		By("Waiting for API server to be ready")
-		root.EventuallyAPIServerReady().Should(Succeed())
-		time.Sleep(time.Second * 5)
+			By("Waiting for API server to be ready")
+			root.EventuallyAPIServerReady().Should(Succeed())
+			time.Sleep(time.Second * 5)
+		}
 	})
 
 	AfterEach(func() {
-		close(stopCh)
-		os.RemoveAll(filepath.Join("/tmp", "indices"))
+		if !f.SelfHostedOperator {
+			close(stopCh)
+			os.RemoveAll(filepath.Join("/tmp", "indices"))
+		}
 	})
 
 	Describe("Search object", func() {
@@ -67,6 +77,7 @@ var _ = Describe("API server", func() {
 				By("Creating deployment: " + deployment.Name)
 				_, err := f.CreateDeployment(*deployment)
 				Expect(err).NotTo(HaveOccurred())
+				f.WaitUntilDeploymentReady(deployment.ObjectMeta)
 
 				// give some time for indexing
 				time.Sleep(time.Second * 30)
